[Offload] Allow kernel launch with unknown argument layout
The C-style CUDA/HIP kernel launch API does not require information
about the arguments. To launch the kernel the driver needs to accept
this vector of argument pointers, or we need to identify their layout.
For CUDA we simply ask the driver to do the work, for AMD/HSA we do it
by scanning the kernel metadata and arranging the arguments ourselves.